// support/enums/topMenuUrls.ts
export enum TopMenuUrl {
    Books = '/books',
    Computers = '/computers',
    Electronics = '/electronics',
    Apparel = '/apparel-shoes',
    DigitalDownloads = '/digital-downloads',
    Jewelry = '/jewelry',
    Gifts = '/gift-cards',
    Desktops = '/desktops',
    Cart = '/cart',
    Wishlist = '/wishlist'
}